fix: tests and refactor api lib
