+++ src/screens/Home.js
@@ -4,7 +4,7 @@ import { useNavigation } from '@react-navigation/native'
 
 import { useSecrets } from '../context/secrets'
 import { useAuthentication } from '../context/authentication'
-import SecretsService from '../lib/secretsService'
+import apiFactory from '../lib/api'
 import routes from '../lib/routeDefinitions'
 import EmptyTokensText from '../components/EmptyTokensText'
 import Actions from '../components/Actions'
@@ -25,25 +25,24 @@ export default function Home() {
   const { secrets, update, remove } = useSecrets()
   const expoToken = usePushToken()
 
-  // TODO: probably needs a refactor, a Class here isn't much useful
-  const secretsService = useMemo(() => new SecretsService(user), [user])
+  const api = useMemo(() => apiFactory({ idToken: user.idToken }), [user])
 
   useEffect(() => {
-    if (!user) return
+    if (!user || !expoToken) return
 
     const register = async () => {
-      await secretsService.registerSubscription({
+      await api.registerSubscription({
         token: expoToken,
         endpoint: 'http://dummy.com', // TODO: dummy endpoint, till backend is updated to make it optional
       })
     }
 
     register()
-  }, [user, secretsService, expoToken])
+  }, [user, api, expoToken])
 
   const handleGenerateToken = async secret => {
     try {
-      const token = await secretsService.generateToken(secret)
+      const token = await api.generateToken(secret)
       await update({ ...secret, token })
     } catch (err) {
       console.log(err)
@@ -52,7 +51,7 @@ export default function Home() {
 
   const handleRevokeToken = async secret => {
     try {
-      const token = await secretsService.revokeToken(secret)
+      const token = await api.revokeToken(secret)
       await update({ ...secret, token })
     } catch (err) {
       console.log(err)
@@ -61,7 +60,7 @@ export default function Home() {
 
   const handleDeleteSecret = async secret => {
     try {
-      await secretsService.revokeToken(secret)
+      await api.revokeToken(secret)
       await remove(secret)
     } catch (err) {
       console.log(err)
